backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
motor==3.6.0
pymongo==4.9.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
